1st pack of fixes

+++ app/src/main/java/orc/zdertis420/playlistmaker/domain/interactor/PlayerInteractor.kt
@@ -15,5 +15,5 @@ interface PlayerInteractor {
 
     fun isPlaying(): Boolean
 
-    fun getCurrentPosition(): Int
+    fun getCurrentPosition(): Long
 }
